Rearranging core-power macros.
