+++ src/module/macros.ts
@@ -7,8 +7,6 @@ import type {
   LancerAttackMacroData,
   LancerMacroData,
   LancerReactionMacroData,
-  LancerStatMacroData,
-  LancerTextMacroData,
 } from "./interfaces";
 // Import JSON data
 import {
@@ -36,11 +34,11 @@ import { getAutomationOptions } from "./settings";
 
 import { getMacroSpeaker, encodeMacroData, encodedMacroWhitelist, ownedItemFromString } from "./macros/util"
 import { renderMacroTemplate } from "./macros/render"
-import { prepareTextMacro, rollTextMacro } from "./macros/text"
 
 export { encodeMacroData, runEncodedMacro } from "./macros/util"
 export { renderMacroTemplate, renderMacroHTML } from "./macros/render"
 export { prepareActivationMacro } from "./macros/activation"
+export { prepareCoreActiveMacro, prepareCorePassiveMacro } from "./macros/core-power"
 export { prepareItemMacro } from "./macros/item"
 export { prepareOverchargeMacro } from "./macros/overcharge"
 export { prepareStatMacro } from "./macros/stat"
@@ -770,74 +768,6 @@ export function rollReactionMacro(actor: LancerActor, data: LancerReactionMacroD
 
   const template = `systems/${game.system.id}/templates/chat/reaction-card.hbs`;
   return renderMacroTemplate(actor, template, data);
-}
-
-/**
- * Prepares a macro to present core active information for
- * @param a     String of the actor ID to roll the macro as, and who we're getting core info for
- */
-export async function prepareCoreActiveMacro(a: string) {
-  // Determine which Actor to speak as
-  let mech = getMacroSpeaker(a);
-  if (!mech || !mech.is_mech()) return;
-
-  var ent = await mech.data.data.derived.mm_promise;
-  if (!ent.Frame) return;
-
-  if (!ent.CurrentCoreEnergy) {
-    ui.notifications!.warn(`No core power remaining on this frame!`);
-    return;
-  }
-
-  let mData: LancerTextMacroData = {
-    title: ent.Frame.CoreSystem.ActiveName,
-    description: ent.Frame.CoreSystem.ActiveEffect,
-    tags: ent.Frame.CoreSystem.Tags,
-  };
-
-  // TODO--setting for this?
-  new Dialog({
-    title: "Consume Core Power?",
-    content: "Consume your mech's core power?",
-    buttons: {
-      submit: {
-        icon: '<i class="fas fa-check"></i>',
-        label: "Yes",
-        callback: async _dlg => {
-          mech?.update({ "data.core_energy": Math.max(ent.CurrentCoreEnergy - 1, 0) });
-          console.log(`Automatically consumed core power for ${ent.LID}`);
-          if (mech) rollTextMacro(mech, mData);
-        },
-      },
-      cancel: {
-        icon: '<i class="fas fa-times"></i>',
-        label: "No",
-      },
-    },
-    default: "submit",
-  }).render(true);
-}
-
-/**
- * Prepares a macro to present core passive information for
- * Checks whether they have a passive since that could get removed on swap
- * @param a     String of the actor ID to roll the macro as, and who we're getting core info for
- */
-export async function prepareCorePassiveMacro(a: string) {
-  // Determine which Actor to speak as
-  let mech = getMacroSpeaker(a);
-  if (!mech || !mech.is_mech()) return;
-
-  var ent = await mech.data.data.derived.mm_promise;
-  if (!ent.Frame) return;
-
-  let mData: LancerTextMacroData = {
-    title: ent.Frame.CoreSystem.PassiveName,
-    description: ent.Frame.CoreSystem.PassiveEffect,
-    tags: ent.Frame.CoreSystem.Tags,
-  };
-
-  rollTextMacro(mech, mData).then();
 }
 
 export async function prepareChargeMacro(a: string | LancerActor) {
